+++ backend/project-board/src/main/java/de/adesso/projectboard/core/rest/application/UserProjectApplicationController.java
@@ -1,4 +1,4 @@
-package de.adesso.projectboard.core.rest;
+package de.adesso.projectboard.core.rest.application;
 
 import de.adesso.projectboard.core.base.rest.application.persistence.ProjectApplicationLog;
 import de.adesso.projectboard.core.base.rest.application.persistence.ProjectApplicationLogRepository;
